requirements.txt
pyshark
requests
python-dotenv
